fix: fix build issue
